Address swiftlint disable comments
- Move arguments of step runner init
- split tests into files per step
